requirements.txt
Django==2.2.7
django-model-utils==3.2.0
Pillow==6.2.1
django-registration==3.0.1
celery==4.3.0
pytz==2019.3
django-extensions==2.2.5
django-nose==1.4.6
nose==1.3.7
loremipsum==1.0.5
django-autoslug==1.9.6
django-test-without-migrations==0.6
PyYAML==5.2
-e git+https://github.com/mysociety/django-popolo.git#egg=django-popolo
django-taggit==1.2.0
django-markdown-deux==1.0.5
-e git+https://github.com/lfalvarez/django-candidator.git@django2#egg=django-candidator
django-picklefield==2.1.1
django-constance==2.6.0
redis==3.5.0
django-redis==4.11.0
django-haystack==2.8.1
django-filter==2.2.0
sorl_thumbnail==12.6.3
django-bootstrap3==12.1.0
django-appconf==1.0.4
django-compressor==2.4
django-sass-processor==0.8
libsass==0.20.0
rcssmin==1.0.6
rjsmin==1.1.0
django-choices==1.7.1
django-hitcount==1.3.2
django_formtools==2.2
-e git+https://github.com/liberation/django-ogp.git#egg=django_ogp
numpy==1.18.4
pybars3==0.9.7
social-auth-app-django==3.1.0
social-auth-core==3.3.3
celery==4.4.2
django-celery-results==1.2.1
django-el-pagination==3.3.0
djangorestframework==3.11.0
pytimeparse==1.1.8